+++ example/lib/main.dart
@@ -17,7 +17,7 @@ class EvalExample extends StatelessWidget {
               import 'package:flutter/material.dart';
 
               class MyApp extends StatelessWidget {
-                const MyApp({Key? key}) : super(key: key);
+                const MyApp({super.key});
 
                 // This widget is the root of your application.
                 @override