+++ lib/src/material/text_field.dart
@@ -1,26 +1,33 @@
 import 'package:dart_eval/dart_eval.dart';
 import 'package:dart_eval/dart_eval_bridge.dart';
+import 'package:dart_eval/stdlib/core.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_eval/src/painting/text_style.dart';
 import 'package:flutter_eval/src/widgets/editable_text.dart';
 import 'package:flutter_eval/src/widgets/framework.dart';
 
 class $TextField implements $Instance {
   static const $type = BridgeTypeRef.spec(BridgeTypeSpec('package:flutter/src/material/text_field.dart', 'TextField'));
 
-  static const $declaration =
-      BridgeClassDef(BridgeClassType($type, isAbstract: false, $extends: $StatefulWidget$bridge.$type),
-          constructors: {
-            '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
-              BridgeParameter('controller', BridgeTypeAnnotation($TextEditingController.$type), true),
-              BridgeParameter('enabled', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.boolType)), true),
-              BridgeParameter('onChanged', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.functionType)), true),
-            ]))
-          },
-          methods: {},
-          getters: {},
-          setters: {},
-          fields: {},
-          wrap: true);
+  static const $declaration = BridgeClassDef(
+      BridgeClassType($type, isAbstract: false, $extends: $StatefulWidget$bridge.$type),
+      constructors: {
+        '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
+          BridgeParameter('controller', BridgeTypeAnnotation($TextEditingController.$type, nullable: true), true),
+          BridgeParameter(
+              'enabled', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.boolType), nullable: true), true),
+          BridgeParameter('style', BridgeTypeAnnotation($TextStyle.$type, nullable: true), true),
+          BridgeParameter(
+              'onChanged', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.functionType), nullable: true), true),
+          BridgeParameter(
+              'onSubmitted', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.functionType), nullable: true), true),
+        ]))
+      },
+      methods: {},
+      getters: {},
+      setters: {},
+      fields: {},
+      wrap: true);
 
   $TextField.wrap(this.$value);
 
@@ -28,7 +35,9 @@ class $TextField implements $Instance {
     return $TextField.wrap(TextField(
       controller: args[0]?.$value,
       enabled: args[1]?.$value,
-      onChanged: args[2]?.$value,
+      style: args[2]?.$value,
+      onChanged: args[3] == null ? null : (value) => (args[3]! as EvalCallable).call(runtime, null, [$String(value)]),
+      onSubmitted: args[4] == null ? null : (value) => (args[4]! as EvalCallable).call(runtime, null, [$String(value)]),
     ));
   }