+++ lib/src/painting.dart
@@ -3,6 +3,10 @@ library painting;
 
 export 'package:flutter/src/painting/alignment.dart';
 export 'package:flutter/src/painting/basic_types.dart';
+export 'package:flutter/src/painting/borders.dart';
+export 'package:flutter/src/painting/box_border.dart';
+export 'package:flutter/src/painting/box_decoration.dart';
+export 'package:flutter/src/painting/decoration.dart';
 export 'package:flutter/src/painting/edge_insets.dart';
 export 'package:flutter/src/painting/text_style.dart';
 ''';