+++ lib/src/widgets/container.dart
@@ -4,6 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_eval/src/foundation/key.dart';
 import 'package:flutter_eval/src/painting/alignment.dart';
+import 'package:flutter_eval/src/painting/decoration.dart';
 import 'package:flutter_eval/src/painting/edge_insets.dart';
 import 'package:flutter_eval/src/rendering/box.dart';
 import 'package:flutter_eval/src/sky_engine/ui/painting.dart';
@@ -23,6 +24,8 @@ class $Container implements Container, $Instance {
           BridgeParameter('alignment', BridgeTypeAnnotation($AlignmentGeometry.$type, nullable: true), true),
           BridgeParameter('padding', BridgeTypeAnnotation($EdgeInsetsGeometry.$type, nullable: true), true),
           BridgeParameter('color', BridgeTypeAnnotation($Color.$type, nullable: true), true),
+          BridgeParameter('decoration', BridgeTypeAnnotation($Decoration.$type, nullable: true), true),
+          BridgeParameter('foregroundDecoration', BridgeTypeAnnotation($Decoration.$type, nullable: true), true),
           BridgeParameter(
               'width', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.doubleType), nullable: true), true),
           BridgeParameter(
@@ -45,12 +48,14 @@ class $Container implements Container, $Instance {
       alignment: args[1]?.$value,
       padding: args[2]?.$value,
       color: args[3]?.$value,
-      width: args[4]?.$value,
-      height: args[5]?.$value,
-      constraints: args[6]?.$value,
-      margin: args[7]?.$value,
-      transformAlignment: args[8]?.$value,
-      child: args[9]?.$value,
+      decoration: args[4]?.$value,
+      foregroundDecoration: args[5]?.$value,
+      width: args[6]?.$value,
+      height: args[7]?.$value,
+      constraints: args[8]?.$value,
+      margin: args[9]?.$value,
+      transformAlignment: args[10]?.$value,
+      child: args[11]?.$value,
     ));
   }